refactor(data-generation): Move from hard-coded to dynamic input
In previous runs of data, we had dataset of values that were:
100,
1,000,
10,000,
...

We then tried to move to a dataset of randomized data with a 1,000 unit (n=1,000) increment.
For example:
1,000
2,000
3,000
4,000
5,000

etc.

However, as we began testing more and more algorithms, it was obvious that it wasn't possible to precaculate the datasets in advance. This was alread placing a large amount of unnecessary files in Git. But, more importantly, differnet algorithms need very different ranges of input to test accurately.

Very poor algorithms (e.g., O(n^3)) were idea for 100, 1,000, 10,000 datasets. Other algorithms (e.g, BubbleSort) needed a set number of values in even increments of 1,000.

However, even moderate algorithms (e.g., O(n)) were so efficient, they couldn't easily be measured on low number of inputs.

This testing framework allows us to dynamically choose the starting number of inputs (e.g., n=1,000), the ending number of inputs (e.g., n=100,000), and the resolution (e.g., every 1,000. n=1,000, 2,000, 3,000, etc.).